signer_service/requirements.txt
# signer_service dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
web3>=6.0.0
eth-account>=0.8.0
